+++ app/projects/devcrate/ngx-dc-data-sources/src/public-api.ts
@@ -5,3 +5,4 @@
 export * from './lib/base.data-source';
 export * from './lib/api.data-source';
 export * from './lib/infinite-scroller.data-source';
+export * from './lib/utils.data-source'