+++ app/projects/devcrate/ngx-dc-dropdown/src/lib/dropdown.component.ts
@@ -1,19 +1,21 @@
 import { CdkFixedSizeVirtualScroll, CdkVirtualForOf, CdkVirtualScrollViewport } from "@angular/cdk/scrolling";
 import { AsyncPipe, NgClass, NgTemplateOutlet } from "@angular/common";
-import { Component, ElementRef, HostBinding, HostListener, OnDestroy, OnInit, ViewEncapsulation, inject, input, output, contentChild, signal, model } from '@angular/core';
+import { Component, ElementRef, HostBinding, HostListener, OnDestroy, OnInit, ViewEncapsulation, contentChild, inject, input, model, output } from '@angular/core';
 import { MatRipple } from "@angular/material/core";
 import { MatIcon } from "@angular/material/icon";
 import { MatListItem, MatListSubheaderCssMatStyler, MatNavList } from "@angular/material/list";
 import { MatProgressBar } from "@angular/material/progress-bar";
-import { NgxDcFilledCountPipe, NgxDcModalService, NgxDcModalServiceToken, NgxDcRolesServiceToken } from "@devcrate/ngx-dc-utils";
+import { NgxDcFilledCountPipe, NgxDcModalService, NgxDcModalServiceToken } from "@devcrate/ngx-dc-utils";
 import { BehaviorSubject } from "rxjs";
 import { takeUntil } from "rxjs/operators";
+import { v4 as uuid } from 'uuid';
 import {
   NgxDcDropdownDataSource
 } from "./dropdown-data.source";
 import {
   NgxDcDropdownHeaderDirective,
-  NgxDcDropdownItemDirective, NgxDcDropdownLoadingDirective, NgxDcDropdownNoItemsDirective
+  NgxDcDropdownItemDirective, NgxDcDropdownLoadingDirective, NgxDcDropdownNoItemsDirective,
+  NgxDcDropdownOptionsHeaderDirective
 } from "./dropdown.directives";
 
 @Component({
@@ -31,20 +33,26 @@ import {
         CdkFixedSizeVirtualScroll,
         CdkVirtualForOf,
         CdkVirtualScrollViewport,
-        MatRipple
+        MatRipple,
     ],
     templateUrl: './dropdown.component.html',
     styleUrl: './dropdown.component.scss',
     encapsulation: ViewEncapsulation.None,
 })
 export class NgxDcDropdownComponent<GetDataItemsT, FinalDataItemsT = GetDataItemsT, RetrievedItemT = GetDataItemsT> implements OnInit, OnDestroy {
+  public dropdownId = `dc-${uuid()}`
   public modalService = inject<NgxDcModalService>(NgxDcModalServiceToken);
 
   // We should put the InfiniteSidePaneListDataSource here because it extends all the other functionality for the data source
   public dataSource = input.required<NgxDcDropdownDataSource<GetDataItemsT, FinalDataItemsT, RetrievedItemT>>();
-
+  
   public opened = model<boolean>(false);
 
+  public readonly allowClickOutside = input<boolean>(true);
+  public readonly allowOptionClick = input<boolean>(true);
+  public optionClicked = output<FinalDataItemsT>();
+  public headerClicked = output<void>();
+  
   public readonly useGlobalLoader = input(false);
   public readonly appearance = input<'fill' | 'outline' | 'none' | 'rounded-fill' | 'rounded-outline'>('rounded-fill');
   public readonly roundedEdgeSize = input<string>('4px');
@@ -61,8 +69,22 @@ export class NgxDcDropdownComponent<GetDataItemsT, FinalDataItemsT = GetDataItem
   @HostListener("document:click", ["$event"])
   public onClick(event: MouseEvent) {
     if (!this.opened()) { return }
-    if (!(event.target as HTMLElement).closest(".dc-dropdown")) {
+    if (!this.allowClickOutside()) { return }
+    const meDropdown = (event.target as HTMLElement).closest(`#${this.dropdownId}`)
+    const clickedOnDropdown = !!meDropdown
+    const anyPanel = Array.from(document.querySelectorAll('.cdk-overlay-pane'))
+    const somePanelIsOpened = anyPanel.length > 0
+    const isInPanel = !!anyPanel.find(panel => panel.querySelector((`#${this.dropdownId}`)))
 
+    // The following shows each scenario
+    // if the dropdown is in a modal, and we didn't click on the dropdown itself, close it
+    // If the dropdown is NOT in the modal, but under the modal, don't close it
+    // If the dropdown is in the modal, and we clicked on the dropdown itself, close it
+    // If the dropdown is NOT in the modal, but under the modal and we clicked on the dropdown itself, don't close it because it's not interactable
+    // if there is no modal opened, and we clicked on the dropdown itself, close it
+    // if there is no modal and we didn't click on the dropdown itself, close it
+
+    if ((isInPanel || !somePanelIsOpened) && !clickedOnDropdown) {
       this.opened.set(false)
     }
   }
@@ -99,6 +121,7 @@ export class NgxDcDropdownComponent<GetDataItemsT, FinalDataItemsT = GetDataItem
 
   public destroy$ = new BehaviorSubject(false)
   readonly dropdownHeader = contentChild(NgxDcDropdownHeaderDirective);
+  readonly dropdownOptionsHeader = contentChild(NgxDcDropdownOptionsHeaderDirective);
 
   readonly dropdownOption = contentChild(NgxDcDropdownItemDirective);
 
@@ -165,15 +188,23 @@ export class NgxDcDropdownComponent<GetDataItemsT, FinalDataItemsT = GetDataItem
     this.opened.set(false)
   }
 
-  public async onPaneItemClicked(item: FinalDataItemsT): Promise<RetrievedItemT> {
+  public async onPaneItemClicked(item: FinalDataItemsT): Promise<void> {
+    if (this.allowOptionClick()) {
+      this.handlePaneItemClick(item)
+    }
+    this.optionClicked.emit(item)
+  }
+
+  public async handlePaneItemClick(item: FinalDataItemsT): Promise<void> {
     await this.dataSource().onPaneItemClicked(item)
     this.opened.set(false)
-    return this.dataSource().retrievedItem.value
   }
 
   protected readonly focus = focus;
 
   public onDropdownClick() {
+    this.headerClicked.emit()
+    if (!this.allowOptionClick()) { return }
     this.opened.set(!this.opened())
   }
 }