+++ app/projects/devcrate/ngx-dc-dropdown/src/lib/dropdown.module.ts
@@ -1,11 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { NgxDcDropdownComponent } from './dropdown.component';
-import { NgxDcDropdownHeaderDirective, NgxDcDropdownItemDirective, NgxDcDropdownLoadingDirective, NgxDcDropdownNoItemsDirective } from './dropdown.directives';
+import { NgxDcDropdownHeaderDirective, NgxDcDropdownItemDirective, NgxDcDropdownLoadingDirective, NgxDcDropdownNoItemsDirective, NgxDcDropdownOptionsHeaderDirective } from './dropdown.directives';
 
 const declarations = [
   NgxDcDropdownComponent,
   NgxDcDropdownHeaderDirective,
+  NgxDcDropdownOptionsHeaderDirective,
   NgxDcDropdownItemDirective,
   NgxDcDropdownLoadingDirective,
   NgxDcDropdownNoItemsDirective,