+++ app/src/app/pages/dropdown-demo/basic-dropdown/basic-dropdown.component.ts
@@ -3,7 +3,8 @@ import { Component, OnInit, inject } from '@angular/core';
 import { FormsModule } from "@angular/forms";
 import { MatSelectModule } from "@angular/material/select";
 import {NgxDcDropdownModule } from "@devcrate/ngx-dc-dropdown";
-import { BasicDropdownDataSource } from "./basic-dropdown.data-source";
+import { BasicDropdownDataSource, InfiniteDropdownDataSource } from "./basic-dropdown.data-source";
+import { NgxDcModalService } from "@devcrate/ngx-dc-utils"
 
 @Component({
     selector: 'basic-dropdown',
@@ -15,13 +16,22 @@ import { BasicDropdownDataSource } from "./basic-dropdown.data-source";
         MatSelectModule,
         FormsModule,
     ],
-    providers: [BasicDropdownDataSource]
+    providers: [BasicDropdownDataSource, InfiniteDropdownDataSource]
 })
 
 export class BasicDropdownComponent implements OnInit {
   dataSource = inject(BasicDropdownDataSource);
+  infiniteDataSource = inject(InfiniteDropdownDataSource)
+  modalService = inject(NgxDcModalService);
 
   public async ngOnInit() {
     await this.dataSource.initialize()
+    this.infiniteDataSource.pageSize = 10
+    this.infiniteDataSource.initialize()
+  }
+  
+  public async refresh() {
+    this.infiniteDataSource.resetInfiniteScrolling()
+    await this.infiniteDataSource.refresh()
   }
 }