+++ app/src/app/pages/mat-select-data-source-demo/basic-select-data-source/basic-select-data-source.component.ts
@@ -1,25 +1,11 @@
+import { JsonPipe } from "@angular/common";
 import { Component, OnInit, inject } from '@angular/core';
-import {BasicSelectDataSourceDataSource} from "./basic-select-data-source.data-source";
-import {
-  NgxDcPaneContentDirective,
-  NgxDcPanePageInfoDirective,
-  NgxDcSidePaneItemDirective,
-  NgxDcSidePaneListComponent
-} from "@devcrate/ngx-dc-side-pane-list";
-import {MatCard, MatCardTitle} from "@angular/material/card";
-import {JsonPipe} from "@angular/common";
-import {
-  NgxDcDropdownComponent,
-  NgxDcDropdownHeaderDirective,
-  NgxDcDropdownItemDirective,
-  NgxDcDropdownLoadingDirective,
-  NgxDcDropdownNoItemsDirective
-} from "@devcrate/ngx-dc-dropdown";
-import {MatSelectModule} from "@angular/material/select";
-import {FormsModule} from "@angular/forms";
+import { FormsModule } from "@angular/forms";
+import { MatSelectModule } from "@angular/material/select";
 import {
   NgxDcMatSelectDataSourceModule
 } from "@devcrate/ngx-dc-mat-select-data-source";
+import { BasicSelectDataSourceDataSource } from "./basic-select-data-source.data-source";
 
 @Component({
     selector: 'basic-select-data-source',