+++ packages/src/eslint-config/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.0.9 (2025-04-10)
+
+
+### Bug Fixes
+
+* scrolling issues ([c5d3097](https://github.com/danda-panda-bytes/devcrate/commit/c5d30977188ad8069ae89035d9f856f572b8aa9b))
+
+
+
+
+
+## 2.0.8 (2025-03-20)
+
+**Note:** Version bump only for package @devcrate/eslint-config
+
+
+
+
+
+## 2.0.7 (2025-03-20)
+
+**Note:** Version bump only for package @devcrate/eslint-config
+
+
+
+
+
+## 2.0.6 (2025-03-20)
+
+**Note:** Version bump only for package @devcrate/eslint-config
+
+
+
+
+
+## 2.0.5 (2025-03-20)
+
+**Note:** Version bump only for package @devcrate/eslint-config
+
+
+
+
+
+## 2.0.4 (2025-03-20)
+
+
+### Bug Fixes
+
+* improve dropdown to have header ([b3a276c](https://github.com/danda-panda-bytes/devcrate/commit/b3a276cd0712762ba3afda03db71030a547dd935))
+
+
+
+
+
 ## 2.0.3 (2025-03-03)
 
 **Note:** Version bump only for package @devcrate/eslint-config